+++ src/components/Template/Template.tsx
@@ -1,59 +1,21 @@
-import { Box, chakra, useStyleConfig, ChakraComponent } from "@chakra-ui/react";
-import React, { forwardRef } from "react";
+import { Box, chakra, ChakraComponent, useStyleConfig } from "@chakra-ui/react";
+import { forwardRef } from "react";
 
-import SkipNavigation from "../SkipNavigation/SkipNavigation";
+export const sidebarPlacementArray = ["none", "left", "right"] as const;
+export type SidebarPlacement = typeof sidebarPlacementArray[number];
 
 export interface TemplateProps {
   /** ID that other components can cross reference for accessibility purposes. */
   id?: string;
-}
-export interface TemplateHeaderProps {
-  /** Flag to render an HTML header element. True by default. */
-  renderHeaderElement?: boolean;
-}
-export interface TemplateFooterProps {
-  /** Flag to render an HTML footer element. True by default. */
-  renderFooterElement?: boolean;
-}
-export interface TemplateSidebarProps {
-  /** Renders the `TemplateContentSidebar` component either on the left or
-   * right side of the `TemplateContentPrimary` component. */
-  sidebar?: "none" | "left" | "right";
-}
-export interface TemplateContentProps extends TemplateSidebarProps {
-  /** ID used for the `main` HTML element. Defaults to "mainContent". Useful
-   * anchor for the application skip navigation. */
-  id?: string;
+  /** Renders the `TemplateSidebar` component either on the left or
+   * right side of the `TemplateMain` component. */
+  sidebar?: SidebarPlacement;
 }
 
-export interface TemplateAppContainerProps
-  extends TemplateFooterProps,
-    TemplateHeaderProps,
-    TemplateSidebarProps {
-  /** DOM that will be rendered before the rest of the components in
-   * `TemplateAppContainer` and immediately before the `TemplateHeader` component. */
-  aboveHeader?: React.ReactElement;
-  /** DOM that will be rendered in the `TemplateBreakout` component section. */
-  breakout?: React.ReactElement;
-  /** DOM that will be rendered in the `TemplateContentBottom` component section. */
-  contentBottom?: React.ReactElement;
+export interface TemplateMainProps {
   /** ID used for the `main` HTML element. Defaults to "mainContent". Useful
    * anchor for the application skip navigation. */
-  contentId?: string;
-  /** DOM that will be rendered in the `TemplateContentPrimary` component section. */
-  contentPrimary?: React.ReactElement;
-  /** DOM that will be rendered in the `TemplateContentSidebar` component section. */
-  contentSidebar?: React.ReactElement;
-  /** DOM that will be rendered in the `TemplateContentTop` component section. */
-  contentTop?: React.ReactElement;
-  /** DOM that will be rendered in the `TemplateFooter` component section. */
-  footer?: React.ReactElement;
-  /** DOM that will be rendered in the `TemplateHeader` component section. */
-  header?: React.ReactElement;
-  /** ID that other components can cross reference for accessibility purposes. */
   id?: string;
-  /** Render the `SkipNavigation` component or not. False by default. */
-  renderSkipNavigation?: boolean;
 }
 
 /**
@@ -67,11 +29,12 @@ const Template: ChakraComponent<
   React.PropsWithChildren<TemplateProps>
 > = chakra(
   forwardRef<HTMLDivElement, React.PropsWithChildren<TemplateProps>>(
-    (props, ref?) => {
-      const styles = useStyleConfig("Template", {});
+    ({ children, sidebar = "none" }, ref?) => {
+      const styles = useStyleConfig("Template", { variant: sidebar });
+
       return (
-        <Box id={props.id} ref={ref} __css={styles} {...props}>
-          {props.children}
+        <Box __css={styles} ref={ref}>
+          {children}
         </Box>
       );
     }
@@ -79,307 +42,87 @@ const Template: ChakraComponent<
 );
 
 /**
- * This optional component renders its children from edge-to-edge and should
- * be used for alerts or notifications that are typically site-wide. This must
- * be rendered immediately before the `TemplateHeader` component. This is meant
- * for components that render an `aside` HTML element or HTML element with the
- * `role="complementary"` attribute. These elements should *not* be rendered
- * in the `header` HTML section since that's an accessibility violation.
+ * This optional component renders its children above the main content
+ * and spans edge-to-edge. It is most useful for `Breadcrumbs`, `Hero`,
+ * or other banner-like components.
  */
-const TemplateAboveHeader: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateProps>
-) => {
+const TemplateBreakout: React.FC<React.PropsWithChildren> = ({ children }) => {
   const styles = useStyleConfig("TemplateBreakout", {});
-  return <Box __css={styles}>{props.children}</Box>;
-};
-
-/**
- * This optional component should be the first child of the `Template`
- * component. This is rendered as an HTML `<header>` element. If an HTML
- * `<header>` element is already passed in a custom component as the children,
- * set `renderFooterElement` to `false`. Otherwise, the parent wrapper will
- * render an HTML `<header>` element.
- */
-const TemplateHeader: React.FC<any> = ({
-  children,
-  renderHeaderElement = true,
-}: React.PropsWithChildren<TemplateHeaderProps>) => {
-  const styles = useStyleConfig("TemplateBreakout", {});
-  let headerElement = <Box __css={styles}>{children}</Box>;
-
-  // The user wants to render the `header` HTML element.
-  if (renderHeaderElement) {
-    // But if they passed in a component that renders an HTML `<header>`,
-    // then log a warning.
-    React.Children.map(children as JSX.Element, (child: React.ReactElement) => {
-      if (child?.type === "header" || child?.props?.mdxType === "header") {
-        console.warn(
-          "NYPL Reservoir TemplateHeader: An HTML `header` element was passed " +
-            "in. Set `renderHeaderElement` to `false` to avoid nested HTML " +
-            "`header` elements."
-        );
-      }
-    });
-    headerElement = (
-      <Box as="header" __css={styles}>
-        {children}
-      </Box>
-    );
-  }
-  return headerElement;
-};
-
-/**
- * This component should be used inside the `Template` component to contain both
- * the `TemplateAboveHeader` and `TemplateHeader` components. This is meant to
- * render its children from edge to edge and is most useful for the headers,
- * `Breadcrumbs`, and `Hero` components or other banner-like components.
- */
-const TemplateBreakout: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateProps>
-) => {
-  const styles = useStyleConfig("TemplateBreakout", {});
-  return <Box __css={styles}>{props.children}</Box>;
-};
-
-/**
- * This component is most useful to render content on the page. This renders an
- * HTML `<main>` element with an id of "mainContent". The "mainContent" id should
- * be used as the consuming application's skip navigation link. The `TemplateContent`
- * component also takes a `sidebar` prop with optional "left" or "right" values.
- * This will set the correct *styling* needed for the `TemplateContentPrimary`
- * and `TemplateContentSidebar` components. Note that `TemplateContentPrimary`
- * and `TemplateContentSidebar` must be ordered correctly as children elements
- * for the appropriate styles to take effect.
- */
-const TemplateContent: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateContentProps>
-) => {
-  const { children, id = "mainContent", sidebar = "none" } = props;
-  const styles = useStyleConfig("TemplateContent", {
-    variant: sidebar,
-  });
-  // Manually pass in the `sidebar` prop to the `TemplateContentPrimary` and
-  // `TemplateContentSidebar` components.
-  const newChildren = React.Children.map(
-    children as JSX.Element,
-    (child: React.ReactElement) => {
-      let newChild = child;
-      if (
-        (child && child?.type === TemplateContentPrimary) ||
-        (child?.props && child.props?.mdxType === "TemplateContentPrimary") ||
-        child?.type === TemplateContentSidebar ||
-        (child?.props && child.props?.mdxType === "TemplateContentSidebar")
-      ) {
-        newChild = React.cloneElement(child, { sidebar });
-      }
-
-      return newChild;
-    }
-  );
 
   return (
-    <Box as="main" id={id} __css={styles}>
-      {newChildren}
+    <Box gridArea="breakout" __css={styles}>
+      {children}
     </Box>
   );
 };
 
 /**
- * This optional component must be used inside the `TemplateContent` component.
- * This renders content in the main width of the container and will always render
- * above the primary component and the sidebar component (if any).
+ * This optional component renders content at a max width of 1280px and
+ * will render below `TemplateBreakout` (if being used) and above the
+ * main content and sidebar (if one exists).
  */
-const TemplateContentTop: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateProps>
-) => {
-  const styles = useStyleConfig("TemplateContentTopBottom", {});
-  return <Box __css={styles}>{props.children}</Box>;
-};
+const TemplateTop: React.FC<React.PropsWithChildren> = ({ children }) => (
+  <Box gridArea="top">{children}</Box>
+);
 
 /**
- * This optional component must be used inside the `TemplateContent` component
- * and after the `TemplateContentPrimary` or `TemplateContentSidebar` component.
- * This renders content in the main width of the container and should always
- * render below the primary component and the sidebar component (if any).
+ * This component renders an HTML `<main>` element with an id of "mainContent".
+ * The "mainContent" id should be used as the consuming application's skip
+ * navigation link. The component should not be used in conjunction with
+ * `TemplateMainNarrow`.
  */
-const TemplateContentBottom: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateProps>
-) => {
-  const styles = useStyleConfig("TemplateContentTopBottom", {});
-  return <Box __css={styles}>{props.children}</Box>;
-};
-
-/**
- * This component is used to render content in a column when there must be a
- * sidebar component on either its left or right side. It must go inside the
- * `TemplateContent` component. An optional `sidebar` prop value of "left" or
- * "right" can be passed to render the correct CSS styles. If the `sidebar`
- * prop is used in the `TemplateContent` component, there is no need to pass
- * the `sidebar` prop to this component -- `TemplateContent` will handle it.
- */
-const TemplateContentPrimary: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateContentProps>
-) => {
-  const { sidebar } = props;
-  const styles = useStyleConfig("TemplateContentPrimary", {
-    variant: sidebar,
-  });
-  return <Box __css={styles}>{props.children}</Box>;
-};
-
-/**
- * This component is used to render content in a sidebar column. It must go
- * inside the `TemplateContent` component and must be paired with the
- * `TemplateContentPrimary` component. If this is a left sidebar, it needs to be
- * rendered before the `TemplateContentPrimary` component. If this is a right
- * sidebar, it needs to be rendered after the `TemplateContentPrimary` component.
- * An optional `sidebar` prop value of "left" or "right" can be passed to render
- * the correct CSS styles. If the `sidebar` prop is used in the `TemplateContent`
- * component, there is no need to pass the `sidebar` prop to this component --
- * `TemplateContent` will handle it.
- */
-const TemplateContentSidebar: React.FC<any> = (
-  props: React.PropsWithChildren<TemplateContentProps>
-) => {
-  const { sidebar } = props;
-  const styles = useStyleConfig("TemplateContentSidebar", {
-    variant: sidebar,
-  });
-  return <Box __css={styles}>{props.children}</Box>;
-};
-
-/**
- * This optional component should be the last child of the `Template`
- * component. This is rendered as an HTML `<footer>` element and spans the full
- * width of the page. If an HTML `<footer>` element is already passed in a
- * custom component, set `renderFooterElement` to `false`.
- */
-const TemplateFooter: React.FC<any> = ({
+const TemplateMain: React.FC<React.PropsWithChildren<TemplateMainProps>> = ({
   children,
-  renderFooterElement = true,
-}: React.PropsWithChildren<TemplateFooterProps>) => {
-  const styles = useStyleConfig("TemplateBreakout", {});
-  let footerElement = <Box __css={styles}>{children}</Box>;
+  id = "mainContent",
+}) => (
+  <Box as="main" id={id} gridArea="main">
+    {children}
+  </Box>
+);
 
-  // The user wants to render the `footer` HTML element.
-  if (renderFooterElement) {
-    // But give a warning if one was passed.
-    React.Children.map(children as JSX.Element, (child: React.ReactElement) => {
-      if (child?.type === "footer" || child?.props?.mdxType === "footer") {
-        console.warn(
-          "NYPL Reservoir TemplateFooter: An HTML `footer` element was passed " +
-            "in. Set `renderFooterElement` to `false` to avoid nested HTML " +
-            "`footer` elements."
-        );
-      }
-    });
-    footerElement = (
-      <Box as="footer" __css={styles}>
+/**
+ * This component renders an HTML `<main>` element with an id of "mainContent".
+ * The "mainContent" id should be used as the consuming application's skip
+ * navigation link. It provides a narrower container for better readability if
+ * the main content includes long text. This component is meant to be used in
+ * lieu of `TemplateMain`, and should not be used in conjunction with
+ * `TemplateSidebar`.
+ */
+const TemplateMainNarrow: React.FC<React.PropsWithChildren<TemplateMainProps>> =
+  ({ children, id = "mainContent" }) => {
+    const styles = useStyleConfig("TemplateMainNarrow");
+
+    return (
+      <Box as="main" id={id} gridArea="main" __css={styles}>
         {children}
       </Box>
     );
-  }
-  return footerElement;
-};
+  };
 
 /**
- * This single component can be used instead of all the individual template
- * components. Instead of importing and rendering the needed "template"
- * components, each section is passed as a prop to the section where it should
- * be rendered. For example, if you want to render content in the
- * `TemplateContentPrimary` section, then pass it as a prop to `contentPrimary`.
+ * This optional component is used to render content in a sidebar column.
+ * It must be paired with the `TemplateMain` component. An optional `sidebar`
+ * prop value of "left" or "right" must be passed to the `Template` wrapper
+ * to render the correct CSS styles.
  */
-export const TemplateAppContainer: ChakraComponent<
-  React.ForwardRefExoticComponent<
-    React.PropsWithChildren<TemplateAppContainerProps> &
-      React.RefAttributes<HTMLDivElement>
-  >,
-  React.PropsWithChildren<TemplateAppContainerProps>
-> = chakra(
-  forwardRef<
-    HTMLDivElement,
-    React.PropsWithChildren<TemplateAppContainerProps>
-  >((props, ref?) => {
-    const {
-      aboveHeader,
-      breakout,
-      contentId = "mainContent",
-      contentBottom,
-      contentPrimary,
-      contentSidebar,
-      contentTop,
-      footer,
-      header,
-      id,
-      sidebar = "none",
-      renderFooterElement = true,
-      renderHeaderElement = true,
-      renderSkipNavigation = false,
-      ...rest
-    } = props;
-    const aboveHeaderElem = aboveHeader && (
-      <TemplateAboveHeader>{aboveHeader}</TemplateAboveHeader>
-    );
-    const contentTopElem = contentTop && (
-      <TemplateContentTop>{contentTop}</TemplateContentTop>
-    );
-    const contentPrimaryElem = contentPrimary && (
-      <TemplateContentPrimary>{contentPrimary}</TemplateContentPrimary>
-    );
-    const contentBottomElem = contentBottom && (
-      <TemplateContentBottom>{contentBottom}</TemplateContentBottom>
-    );
-    const contentSidebarElem = contentSidebar && (
-      <TemplateContentSidebar>{contentSidebar}</TemplateContentSidebar>
-    );
-    return (
-      <>
-        {renderSkipNavigation ? <SkipNavigation /> : null}
-        <Template id={id} ref={ref} {...rest}>
-          <TemplateBreakout>
-            {aboveHeaderElem}
-            {(header || breakout) && (
-              <TemplateHeader renderHeaderElement={renderHeaderElement}>
-                {header}
-                {breakout}
-              </TemplateHeader>
-            )}
-          </TemplateBreakout>
-          {/* Note that setting `sidebar` as a prop here affects the
-          TemplateContentSidebar and TemplateContentPrimary components. */}
-          <TemplateContent id={contentId} sidebar={sidebar}>
-            {contentTopElem}
+const TemplateSidebar: React.FC<React.PropsWithChildren> = ({ children }) => (
+  <Box gridArea="sidebar">{children}</Box>
+);
 
-            {sidebar === "left" && contentSidebarElem}
-
-            {contentPrimaryElem}
-
-            {sidebar === "right" && contentSidebarElem}
-
-            {contentBottomElem}
-          </TemplateContent>
-          {footer && (
-            <TemplateFooter renderFooterElement={renderFooterElement}>
-              {footer}
-            </TemplateFooter>
-          )}
-        </Template>
-      </>
-    );
-  })
+/**
+ * This optional component renders content at a max width of 1280px and
+ * will always render below the main content and sidebar (if one exists).
+ */
+const TemplateBottom: React.FC<React.PropsWithChildren> = ({ children }) => (
+  <Box gridArea="bottom">{children}</Box>
 );
 
 export {
   Template,
-  TemplateAboveHeader,
   TemplateBreakout,
-  TemplateContent,
-  TemplateContentBottom,
-  TemplateContentPrimary,
-  TemplateContentSidebar,
-  TemplateContentTop,
-  TemplateFooter,
-  TemplateHeader,
+  TemplateTop,
+  TemplateMain,
+  TemplateMainNarrow,
+  TemplateSidebar,
+  TemplateBottom,
 };
-export default TemplateAppContainer;